refactor(matcher): separate different matching types from main MatchParser
